# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.CourseProgress:
    properties:
      completed_count:
        type: integer
      course_id:
        type: integer
      course_title:
        type: string
      percentage:
        type: number
      total_lessons:
        type: integer
    type: object
...
  models.CreateLessonRequest:
    properties:
      content:
        minLength: 1
        type: string
      order_index:
        minimum: 0
        type: integer
      title:
        maxLength: 200
        minLength: 1
        type: string
    required:
    - content
    - title
    type: object
  models.CreateSubmissionRequest:
    properties:
      code:
        minLength: 1
        type: string
    required:
    - code
    type: object
...
  models.ErrorBody:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  models.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/models.ErrorBody'
    type: object
  models.Lesson:
    properties:
      content:
        type: string
      course_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      order_index:
        type: integer
      title:
        type: string
      updated_at:
        type: string
    type: object
...
  models.PaginatedResponse:
    properties:
      data: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
...
  models.Submission:
    properties:
      code:
        type: string
      id:
        type: integer
      result:
        type: string
      score:
        type: integer
      status:
        $ref: '#/definitions/models.SubmissionStatus'
      submitted_at:
        type: string
      task_id:
        type: integer
      user_id:
        type: integer
    type: object
  models.SubmissionStatus:
    enum:
    - pending
    - passed
    - failed
    type: string
    x-enum-varnames:
    - StatusPending
    - StatusPassed
    - StatusFailed
...
  models.UpdateLessonRequest:
    properties:
      content:
        minLength: 1
        type: string
      order_index:
        minimum: 0
        type: integer
      title:
        maxLength: 200
        minLength: 1
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Online programming learning platform API
  title: Diplom API
  version: "1.0"
paths:
...
  /courses/{course_id}/lessons:
    get:
      description: Get a paginated list of lessons for a course
      parameters:
      - description: Course ID
        in: path
        name: course_id
        required: true
        type: integer
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PaginatedResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List lessons in a course
      tags:
      - Lessons
    post:
      consumes:
      - application/json
      description: Create a new lesson in a course
      parameters:
      - description: Course ID
        in: path
        name: course_id
        required: true
        type: integer
      - description: Lesson data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.CreateLessonRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Lesson'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a lesson (admin)
      tags:
      - Lessons
  /courses/{course_id}/progress:
    get:
      description: Get the authenticated user's progress for a specific course
      parameters:
      - description: Course ID
        in: path
        name: course_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CourseProgress'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get course progress
      tags:
      - Progress
...
  /lessons/{id}:
    delete:
      description: Delete lesson by ID
      parameters:
      - description: Lesson ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a lesson (admin)
      tags:
      - Lessons
    get:
      description: Get lesson by ID
      parameters:
      - description: Lesson ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Lesson'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a lesson
      tags:
      - Lessons
    put:
      consumes:
      - application/json
      description: Update lesson by ID
      parameters:
      - description: Lesson ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UpdateLessonRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Lesson'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a lesson (admin)
      tags:
      - Lessons
...
  /progress:
    get:
      description: Get the authenticated user's progress across all courses
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.CourseProgress'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all progress
      tags:
      - Progress
  /submissions:
    get:
      description: Get a paginated list of all submissions
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PaginatedResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all submissions (admin)
      tags:
      - Submissions
  /submissions/{id}:
    get:
      description: Get submission by ID
      parameters:
      - description: Submission ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Submission'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a submission
      tags:
      - Submissions
...
  /tasks/{task_id}/submissions:
    get:
      description: Get a paginated list of the authenticated user's submissions for
        a task
      parameters:
      - description: Task ID
        in: path
        name: task_id
        required: true
        type: integer
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PaginatedResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List own submissions for a task
      tags:
      - Submissions
    post:
      consumes:
      - application/json
      description: Submit code for a task
      parameters:
      - description: Task ID
        in: path
        name: task_id
        required: true
        type: integer
      - description: Submission data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.CreateSubmissionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Submission'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Submit a solution
      tags:
      - Submissions
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"